Moved the instance quotas test to their own module, increased the instance count.
* This allows the quota test to work while not using low quotas which would break the pagination tests.
