+++ scripts/conf/test_begin.conf
@@ -36,7 +36,7 @@
     "reddwarf_can_have_volume":true,
     "reddwarf_main_instance_has_volume": false,
     "reddwarf_max_accepted_volume_size": 25,
-    "reddwarf_max_instances_per_user": 5,
+    "reddwarf_max_instances_per_user": 55,
     "use_reaper":false,
     "users": [
         { "auth_user":"radmin",
