+++ tests/integration/tests/api/instances_quotas.py
@@ -1,0 +1,47 @@
+from proboscis import before_class
+from proboscis import test
+from proboscis.asserts import assert_raises
+
+from reddwarfclient import exceptions
+from tests.util import test_config
+from tests.util import create_client
+from tests.util import poll_until
+
+
+@test(groups=['dbaas.api.instances.quotas'])
+class InstanceQuotas(object):
+
+    created_instances = []
+
+    @before_class
+    def setup(self):
+        self.client = create_client(is_admin=False)
+
+    @test
+    def test_too_many_instances(self):
+        self.created_instances = []
+        if 'reddwarf_max_instances_per_user' in test_config.values:
+            too_many = test_config.values['reddwarf_max_instances_per_user']
+            already_there = len(self.client.instances.list())
+            flavor = 1
+            for i in range(too_many - already_there):
+                response = self.client.instances.create('too_many_%d' % i,
+                                                  flavor,
+                                                  {'size': 1})
+            # This one better fail, because we just reached our quota.
+            assert_raises(exceptions.OverLimit,
+                          self.client.instances.create,
+                          "too_many", flavor,
+                          {'size': 1})
+
+    @test(runs_after=[test_too_many_instances])
+    def delete_excessive_entries(self):
+        # Delete all the instances called too_many*.
+        for id in self.created_instances:
+            while True:
+                try:
+                    self.client.instances.delete(id)
+                except exceptions.UnprocessableEntity:
+                    continue
+                except exceptions.NotFound:
+                    break
